requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
duckduckgo_search==6.2.13
urlextract==1.9.0
python-dotenv==1.0.1
email-validator==2.2.0
text-unidecode==1.3
phonenumberslite==8.13.45
lxml==5.3.0
readability-lxml==0.8.1
trafilatura==1.12.2
html5lib==1.1
